Добавил функцию unSaveLock для совместного редактирования (для возможности сбросить lock сохранения) Добавил проверку на корректность deleteIndex (он мог быть больше, чем текущий) fix bug #28967 (при сохранении, если запрос на lock чуть долгий и мы успеваем в этот момент сделать долгую вставку данных, то у нас эти данных пропадают)
git-svn-id: svn://192.168.3.15/activex/AVS/Sources/TeamlabOffice/trunk/nodeJSProjects@61694 954022d7-b5bf-4e40-9824-e11837661b57

+++ CoAuthoring/sources/DocsCoServer.js
@@ -553,6 +553,7 @@ exports.install = function (server, callbackFunction) {
 					case 'getLock'				: getLock(conn, data, false); break;
 					case 'saveChanges'			: saveChanges(conn, data); break;
 					case 'isSaveLock'			: isSaveLock(conn, data); break;
+					case 'unSaveLock'			: unSaveLock(conn, -1); break;	// Индекс отправляем -1, т.к. это экстренное снятие без сохранения
 					case 'getMessages'			: getMessages(conn, data); break;
 					case 'unLockDocument'		: checkEndAuthLock(data.isSave, conn.docId, conn.user.id, null, conn); break;
 				}
@@ -1244,10 +1245,13 @@ exports.install = function (server, callbackFunction) {
 			deleteIndex = data.deleteIndex;
 			if (-1 !== deleteIndex) {
 				var deleteCount = pucker.index - deleteIndex;
-				if (objChangesDocument)
-					objChangesDocument.splice(deleteIndex, deleteCount);
-				pucker.index -= deleteCount;
-				sqlBase.deleteChanges(docId, deleteIndex);
+				if (0 < deleteCount) {
+					if (objChangesDocument)
+						objChangesDocument.splice(deleteIndex, deleteCount);
+					pucker.index -= deleteCount;
+					sqlBase.deleteChanges(docId, deleteIndex);
+				} else
+					logger.error("saveChanges docid: %s ; deleteIndex: %s ; startIndex: %s ; deleteCount: %s", docId, deleteIndex, pucker.index, deleteCount);
 			}
 		}
